+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - feat(networking): Add transit gateway static CIDR blocks and Transit Gateway Connect attachments
 - feat(autoScalingGroup): Add option to set maxInstanceLifetime property for AutoScalingGroups
+- feat(securityHub): Allow SecurityHub to be enabled when AwsConfig is enabled with deploymentTargets option
 
 ### Changed